+++ tagger-lib/src/main/java/com/osacky/tagger/TaggerApiExtension.java
@@ -1,0 +1,10 @@
+package com.osacky.tagger;
+
+interface TaggerApiExtension {
+
+    void tag(String tag);
+
+    void value(String name, String value);
+
+    void link(String name, String url);
+}