+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.3
+* Rename ScanApi to TaggerApi
+* Can now configure `taggerApi` extension instead of calling new TaggerApi(project). This make for a more idiomatic Gradle configuration.
+
 ## 0.2
 * Add support for java 8
 * Change package name